Fix issues serving paths with a query string
A limitation of the SimpleHTTPRequestHandler is that it doesn't handle
query strings well. Requests to /?q=search get a slash appended to the
end as if it were a directory, leaving the user at /?q=search/ which
then in turn changes the value of the query string which we process in
JavaScript.

This change hooks into the request process and throws away the query
string so it doesn't cause a redirect.